requirements.txt
streamlit==1.28.1
groq==0.3.2
python-dotenv==1.0.0
regex==2023.10.3
pandas==2.1.3
streamlit-option-menu==0.3.6
